# tienda-web-carpinteria/target/generated-sources/openapi/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: desarrolloweb@tiendaweb.com
  description: Tienda web de carpinteria
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://www.tiendaweb.com.ar
  title: Tienda de Carpinteria - OpenAPI 3.0
  version: 1.0.0
...
servers:
- url: https://localhost:8080/api/v1
security:
- bearerAuth: []
...
paths:
...
  /products/product:
    get:
      description: Retrieve a list of products.
      operationId: retrieveAllProducts
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseContainerResponseDTO'
          description: successful operation response a ProductListDTO as data
      summary: Retrieve a list of products
      tags:
      - backoffice
      x-accepts: application/json
      x-tags:
      - tag: backoffice
    post:
      description: "Create a product, authentication required."
      operationId: createProduct
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequestDTO'
        description: Created product object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseContainerResponseDTO'
          description: successful operation response a ProductDTO as data
        "401":
          description: Access token is missing or invalid
      security:
      - bearerAuth: []
      summary: Create product
      tags:
      - backoffice
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: backoffice
  /products/product/{productId}:
    delete:
      description: "Delete a product, authentication required."
      operationId: deleteProduct
      parameters:
      - description: the product id
        explode: false
        in: path
        name: productId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseContainerResponseDTO'
          description: successful operation response a EmptyResponseDTO as data
        "401":
          description: Access token is missing or invalid
      security:
      - bearerAuth: []
      summary: Delete product
      tags:
      - backoffice
      x-accepts: application/json
      x-tags:
      - tag: backoffice
    get:
      description: "Retrieve a product, authentication required."
      operationId: retrieveProduct
      parameters:
      - description: the product id
        explode: false
        in: path
        name: productId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseContainerResponseDTO'
          description: successful operation response a ProductDTO as data
      security:
      - bearerAuth: []
      summary: Retrieve product
      tags:
      - backoffice
      x-accepts: application/json
      x-tags:
      - tag: backoffice
    put:
      description: "Update a product, authentication required."
      operationId: updateProduct
      parameters:
      - description: the product id
        explode: false
        in: path
        name: productId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductDTO'
        description: Updated product object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseContainerResponseDTO'
          description: successful operation response a ProductDTO as data
        "401":
          description: Access token is missing or invalid
      security:
      - bearerAuth: []
      summary: Update product
      tags:
      - backoffice
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: backoffice
  /products/category:
    get:
      description: Retrieve a list of category.
      operationId: retrieveAllCategory
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseContainerResponseDTO'
          description: successful operation response a CategoryListDTO as data
      summary: Retrieve a list of category
      tags:
      - backoffice
      x-accepts: application/json
      x-tags:
      - tag: backoffice
    post:
      description: "Create a category, authentication required."
      operationId: createCategory
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryRequestDTO'
        description: Created category object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseContainerResponseDTO'
          description: successful operation response a CategoryDTO as data
        "401":
          description: Access token is missing or invalid
      security:
      - bearerAuth: []
      summary: Create category
      tags:
      - backoffice
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: backoffice
  /products/category/{categoryId}:
    delete:
      description: "Delete a category, authentication required."
      operationId: deleteCategory
      parameters:
      - description: the category id
        explode: false
        in: path
        name: categoryId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseContainerResponseDTO'
          description: successful operation response a EmptyResponseDTO as data
        "401":
          description: Access token is missing or invalid
      security:
      - bearerAuth: []
      summary: Delete category
      tags:
      - backoffice
      x-accepts: application/json
      x-tags:
      - tag: backoffice
    get:
      description: "Retrieve a category, authentication required."
      operationId: retrieveCategory
      parameters:
      - description: the category id
        explode: false
        in: path
        name: categoryId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseContainerResponseDTO'
          description: successful operation response a CategoryDTO as data
      security:
      - bearerAuth: []
      summary: Retrieve category
      tags:
      - backoffice
      x-accepts: application/json
      x-tags:
      - tag: backoffice
    put:
      description: "Update a category, authentication required."
      operationId: updateCategory
      parameters:
      - description: the category id
        explode: false
        in: path
        name: categoryId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryDTO'
        description: Updated category object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseContainerResponseDTO'
          description: successful operation response a CategoryDTO as data
        "401":
          description: Access token is missing or invalid
      security:
      - bearerAuth: []
      summary: Update category
      tags:
      - backoffice
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: backoffice
...
components:
...
  schemas:
...
    UserDTO:
      example:
        firstName: Juan
        lastName: Lopez
        password: p4$sW0rD!
        address: Av. Rivadavia 1000
        dateCreated: 2023-01-01T00:00:00.000+00:00
        dateOfBirthday: 2023-01-01T00:00:00.000+00:00
        dni: 40000000
        email: jlopez@email.com
        dateDeleted: 2023-01-01T00:00:00.000+00:00
      properties:
        dni:
          example: 40000000
          format: int32
          type: integer
        firstName:
          example: Juan
          type: string
        lastName:
          example: Lopez
          type: string
        address:
          example: Av. Rivadavia 1000
          type: string
        email:
          description: this field represent the user email
          example: jlopez@email.com
          format: email
          type: string
        password:
          example: p4$sW0rD!
          maxLength: 16
          minLength: 6
          type: string
        dateOfBirthday:
          example: 2023-01-01
          format: date
          type: string
        dateCreated:
          example: 2023-01-01
          format: date
          type: string
        dateDeleted:
          example: 2023-01-01
          format: date
          type: string
      required:
      - address
      - dateCreated
      - dni
      - email
      - firstName
      - lastName
      - password
      type: object
...
    UserListDTO:
      properties:
        items:
          items:
            $ref: '#/components/schemas/UserDTO'
          type: array
      type: object
    UserRoleDTO:
      properties:
        email:
          description: this field represent the user email
          example: john@email.com
          format: email
          type: string
        role:
          example: ROLE_USER
          type: string
        level:
          example: 1
          type: integer
        dateCreated:
          example: 2023-02-01
          format: date
          type: string
        dateDeleted:
          example: 2023-02-01
          format: date
          type: string
      required:
      - dateCreated
      - level
      - role
      type: object
...
    UserRoleListDTO:
      properties:
        items:
          items:
            $ref: '#/components/schemas/UserRoleDTO'
          type: array
      type: object
    CategoryDTO:
      example:
        dateCreated: 2023-02-01T00:00:00.000+00:00
        name: Cajas
        parentCategoryId: 1
        id: 8
        dateDeleted: 2023-02-01T00:00:00.000+00:00
      properties:
        id:
          example: 8
          format: int64
          type: integer
        name:
          example: Cajas
          type: string
        parentCategoryId:
          example: 1
          format: int64
          type: integer
        dateCreated:
          example: 2023-02-01
          format: date
          type: string
        dateDeleted:
          example: 2023-02-01
          format: date
          type: string
      required:
      - dateCreated
      - name
      type: object
    CategoryListDTO:
      properties:
        items:
          items:
            $ref: '#/components/schemas/CategoryDTO'
          type: array
      type: object
    CategoryRequestDTO:
      example:
        name: Cajas
        parentCategoryId: 1
      properties:
        name:
          example: Cajas
          type: string
        parentCategoryId:
          example: 1
          format: int64
          type: integer
      required:
      - name
      type: object
    ProductDTO:
      example:
        dateCreated: 2023-01-01T00:00:00.000+00:00
        price: 3500
        qty: 2
        description: Caja de te de 4 divisiones
        id: 37
        category:
          dateCreated: 2023-02-01T00:00:00.000+00:00
          name: Cajas
          parentCategoryId: 1
          id: 8
          dateDeleted: 2023-02-01T00:00:00.000+00:00
        dateDeleted: 2023-01-02T00:00:00.000+00:00
      properties:
        id:
          example: 37
          format: int64
          type: integer
        description:
          example: Caja de te de 4 divisiones
          type: string
        price:
          example: 3500
          format: double
          type: number
        qty:
          description: the stock quantity.
          example: 2
          type: integer
        category:
          $ref: '#/components/schemas/CategoryDTO'
        dateCreated:
          example: 2023-01-01
          format: date
          type: string
        dateDeleted:
          example: 2023-01-02
          format: date
          type: string
      required:
      - dateCreated
      - name
      - price
      - qty
      type: object
    ProductRequestDTO:
      example:
        price: 35000
        qty: 2
        description: Caja de te de 4 divisiones
        category:
          dateCreated: 2023-02-01T00:00:00.000+00:00
          name: Cajas
          parentCategoryId: 1
          id: 8
          dateDeleted: 2023-02-01T00:00:00.000+00:00
      properties:
        description:
          example: Caja de te de 4 divisiones
          type: string
        price:
          example: 35000
          format: double
          type: number
        qty:
          description: the stock quantity.
          example: 2
          type: integer
        category:
          $ref: '#/components/schemas/CategoryDTO'
      required:
      - category
      - name
      - price
      - qty
      type: object
    ProductListDTO:
      properties:
        items:
          items:
            $ref: '#/components/schemas/ProductDTO'
          type: array
      type: object
    OrderListDTO:
      properties:
        items:
          items:
            $ref: '#/components/schemas/OrderDTO'
          type: array
      type: object
    OrderDTO:
      properties:
        id:
          example: 10
          format: int64
          type: integer
        user:
          $ref: '#/components/schemas/UserDTO'
        orderList:
          items:
            $ref: '#/components/schemas/OrderDetailsDTO'
          type: array
        price:
          format: double
          type: number
        dateCreated:
          example: 2023-02-01
          format: date
          type: string
        dateDeleted:
          example: 2023-02-01
          format: date
          type: string
      required:
      - id
      - price
      - productList
      - user
      type: object
    OrderRequestDTO:
      example:
        orderDetails:
        - product:
            dateCreated: 2023-01-01T00:00:00.000+00:00
            price: 3500
            qty: 2
            description: Caja de te de 4 divisiones
            id: 37
            category:
              dateCreated: 2023-02-01T00:00:00.000+00:00
              name: Cajas
              parentCategoryId: 1
              id: 8
              dateDeleted: 2023-02-01T00:00:00.000+00:00
            dateDeleted: 2023-01-02T00:00:00.000+00:00
          dateCreated: 2023-02-01T00:00:00.000+00:00
          id: 10
          user:
            firstName: Juan
            lastName: Lopez
            password: p4$sW0rD!
            address: Av. Rivadavia 1000
            dateCreated: 2023-01-01T00:00:00.000+00:00
            dateOfBirthday: 2023-01-01T00:00:00.000+00:00
            dni: 40000000
            email: jlopez@email.com
            dateDeleted: 2023-01-01T00:00:00.000+00:00
          order: null
        - product:
            dateCreated: 2023-01-01T00:00:00.000+00:00
            price: 3500
            qty: 2
            description: Caja de te de 4 divisiones
            id: 37
            category:
              dateCreated: 2023-02-01T00:00:00.000+00:00
              name: Cajas
              parentCategoryId: 1
              id: 8
              dateDeleted: 2023-02-01T00:00:00.000+00:00
            dateDeleted: 2023-01-02T00:00:00.000+00:00
          dateCreated: 2023-02-01T00:00:00.000+00:00
          id: 10
          user:
            firstName: Juan
            lastName: Lopez
            password: p4$sW0rD!
            address: Av. Rivadavia 1000
            dateCreated: 2023-01-01T00:00:00.000+00:00
            dateOfBirthday: 2023-01-01T00:00:00.000+00:00
            dni: 40000000
            email: jlopez@email.com
            dateDeleted: 2023-01-01T00:00:00.000+00:00
          order: null
        id: 10
        user:
          firstName: Juan
          lastName: Lopez
          password: p4$sW0rD!
          address: Av. Rivadavia 1000
          dateCreated: 2023-01-01T00:00:00.000+00:00
          dateOfBirthday: 2023-01-01T00:00:00.000+00:00
          dni: 40000000
          email: jlopez@email.com
          dateDeleted: 2023-01-01T00:00:00.000+00:00
      properties:
        id:
          example: 10
          format: int64
          type: integer
        user:
          $ref: '#/components/schemas/UserDTO'
        orderDetails:
          items:
            $ref: '#/components/schemas/OrderDetailsDTO'
          type: array
      type: object
    OrderDetailsDTO:
      example:
        product:
          dateCreated: 2023-01-01T00:00:00.000+00:00
          price: 3500
          qty: 2
          description: Caja de te de 4 divisiones
          id: 37
          category:
            dateCreated: 2023-02-01T00:00:00.000+00:00
            name: Cajas
            parentCategoryId: 1
            id: 8
            dateDeleted: 2023-02-01T00:00:00.000+00:00
          dateDeleted: 2023-01-02T00:00:00.000+00:00
        dateCreated: 2023-02-01T00:00:00.000+00:00
        id: 10
        user:
          firstName: Juan
          lastName: Lopez
          password: p4$sW0rD!
          address: Av. Rivadavia 1000
          dateCreated: 2023-01-01T00:00:00.000+00:00
          dateOfBirthday: 2023-01-01T00:00:00.000+00:00
          dni: 40000000
          email: jlopez@email.com
          dateDeleted: 2023-01-01T00:00:00.000+00:00
        order: null
      properties:
        id:
          example: 10
          format: int64
          type: integer
        user:
          $ref: '#/components/schemas/UserDTO'
        product:
          $ref: '#/components/schemas/ProductDTO'
        order:
          $ref: '#/components/schemas/OrderRequestDTO'
        dateCreated:
          example: 2023-02-01
          format: date
          type: string
      type: object
    JWTResponseDTO:
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
      type: object
    EmptyResponseDTO:
      properties:
        date:
          format: date-time
          type: string
      type: object
...
    ErrorItemDTO:
      example:
        code: code
        link: link
        detail: detail
      properties:
        code:
          type: string
        link:
          type: string
        detail:
          type: string
      type: object
    MetaInformationResponseDTO:
      example:
        refNumber: refNumber
        time: 0
      properties:
        time:
          format: int64
          type: integer
        refNumber:
          type: string
      required:
      - time
      type: object
    DataResponseDTO:
      discriminator:
        propertyName: type
      example:
        type: type
      oneOf:
      - $ref: '#/components/schemas/JWTResponseDTO'
      - $ref: '#/components/schemas/ProductDTO'
      - $ref: '#/components/schemas/UserListDTO'
      - $ref: '#/components/schemas/UserDTO'
      - $ref: '#/components/schemas/OrderRequestDTO'
      - $ref: '#/components/schemas/OrderDTO'
      - $ref: '#/components/schemas/OrderListDTO'
      - $ref: '#/components/schemas/ProductListDTO'
      - $ref: '#/components/schemas/UserRoleDTO'
      - $ref: '#/components/schemas/UserRoleListDTO'
      - $ref: '#/components/schemas/EmptyResponseDTO'
      - $ref: '#/components/schemas/CategoryListDTO'
      - $ref: '#/components/schemas/CategoryDTO'
      properties:
        type:
          description: The type of the contained object
          type: string
      type: object
      x-one-of-name: DataResponseDTO
    ResponseContainerResponseDTO:
      example:
        data:
          type: type
        meta:
          refNumber: refNumber
          time: 0
        errors:
        - code: code
          link: link
          detail: detail
        - code: code
          link: link
          detail: detail
      properties:
        errors:
          items:
            $ref: '#/components/schemas/ErrorItemDTO'
          type: array
        meta:
          $ref: '#/components/schemas/MetaInformationResponseDTO'
        data:
          $ref: '#/components/schemas/DataResponseDTO'
      required:
      - meta
      - type
      type: object
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http